Significant number of changes. Mostly about object and xref streams.
Added standard fonts object to handle loading the 14 standard fonts and
interface with the font metrics objects.

Added font metrics objects to read the standard 14 font metrics files.
Principally, this is to obtain the character widths. Also added the
manager object to handle managing multiple fonts and import from zip
files (e.g., the standard 14 included in this module).

Added Identity-H and Identity-V cmaps I found in a github repo (but
looks like they're from Adobe originally).

Added SCN and scn commands.

With font metrics, widths can also be a dictionary keyed on character
name. Handle that.

Added missing objects to __all__ as appropriate.

Different PDF writers handle trailers a little differently. Needed to
improve handling of them.

Added object stream and xref stream handling.

Changed LoadObject and GetObject to pass an IndirectObject or a tuple of
object id and generation. Really should not have been two separate
arguments. This becomes more important with the xref stream that adds a
'third' type of xref that is a reference to an object in an object
stream. (This required lots of changes in numerous files to condense two
arguments into one.)

Because of object streams, some parsing functions need to handle a
direct object (whose data is tokens[0].values[2]) or a stream object
that jumps right to the value.

Spruce up CMap builder to handle ranges. This is necessary for the
Identity-H and Identity-V maps.

Add comment handling to CMap tokenizer.

Implemented __setattr__ on PDFHigherBase so that objects can be
modified.